requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
bcrypt==4.1.2
SQLAlchemy==2.0.28
pyodbc==5.0.1
sqlalchemy-pyodbc-mssql==0.1.1
